Remove old RabbitMQ controller + associated types
