+++ libs/adapters/src/rabbitmq/index.ts
@@ -1,11 +1,5 @@
 export { RabbitMQAdapter } from './RabbitMQAdapter';
-export { MockRabbitMQController } from './mockRabbitMQController';
 export * from './rabbitMQConfig';
 export { getRabbitMQConfig } from './rabbitMQConfig';
-export {
-  RabbitMQController,
-  RabbitMQControllerError,
-} from './rabbitMQController';
-export { RepublishFailedMessages } from './republishFailedMessages';
 export * from './types';
 export * from './util';
